refactor: Complete Phase 2 - Split useCanvas into focused hooks
## Summary
Complete the monolithic useCanvas hook refactoring by creating focused, single-responsibility hooks. Reduced main hook from 721 lines to 84 lines while maintaining all functionality.

## Changes Made
- **NEW**: `src/hooks/useCanvasDrawing.ts` - Drawing operations and canvas updates (182 lines)
- **NEW**: `src/hooks/useCanvasHistory.ts` - History management and undo functionality (135 lines)
- **REFACTORED**: `src/hooks/useCanvas.ts` - Now a lightweight orchestrator (84 lines)
- **UPDATED**: `REFACTORING_PLAN.md` - Document Phase 2 completion

## Architecture Achievement
Transformed monolithic 721-line hook into 5 focused hooks:
1. **useCanvasRefs** - Canvas references & contexts (97 lines)
2. **useCanvasEvents** - Mouse/touch event handling (134 lines)
3. **useCanvasDrawing** - Drawing operations & updates (182 lines)
4. **useCanvasHistory** - History management & undo (135 lines)
5. **useCanvas** - Lightweight orchestrator (84 lines)

## Benefits
- **88% reduction** in main hook complexity (721 → 84 lines)
- **Clear separation of concerns** - each hook has single responsibility
- **Improved testability** - each hook can be tested independently
- **Better maintainability** - easier to debug and modify specific functionality
- **Enhanced reusability** - focused hooks can be reused independently

## Testing
- All 5 E2E tests passing ✅
- Drawing functionality preserved (3115 white pixels detected)
- Event handling working correctly (mouse/touch)
- History and undo functionality preserved
- Canvas mirroring still working (folded → unfolded)
- Zero regressions detected
